Fix warning.
git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/branches/syslog_aiwi@2081 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/swre/swre_interface.ss/cpp/swre_interface_cpp/TraceStreamBuffer.cpp
@@ -84,7 +84,7 @@ namespace Internal
         using Safir::Dob::Typesystem::Utilities::ToUtf8;
 
         bool success;
-        SwreC_TraceAppendString(m_prefixId, ToUtf8(std::wstring(s, num)).c_str(), success);
+        SwreC_TraceAppendString(m_prefixId, ToUtf8(std::wstring(s, static_cast<size_t>(num))).c_str(), success);
         if (!success)
         {
             Safir::Dob::Typesystem::LibraryExceptions::Instance().Throw();
